+++ tangelo/algorithms/variational/iqcc_solver.py
@@ -28,7 +28,7 @@ Refs:
         J. Chem. Theory Comput. 2020, 16, 2, 1055–1063.
 """
 
-from tangelo.linq import Simulator
+from tangelo.linq import get_backend
 from tangelo.toolboxes.ansatz_generator.qcc import QCC
 from tangelo.algorithms.variational.vqe_solver import VQESolver, BuiltInAnsatze
 from tangelo.toolboxes.ansatz_generator._qubit_cc import qcc_op_dress
@@ -62,8 +62,7 @@ class iQCC_solver:
             followed by all spin down. Default, False.
         initial_var_params (str or array-like): Initial values of the variational parameters
             for the classical optimizer.
-        backend_options (dict): Parameters to build the tangelo.linq Simulator
-            class.
+        backend_options (dict): parameters to build the underlying compute backend (simulator, etc).
         penalty_terms (dict): Parameters for penalty terms to append to target
             qubit Hamiltonian (see penaly_terms for more details).
         ansatz_options (dict): Parameters for the chosen ansatz (see given ansatz
@@ -156,7 +155,7 @@ class iQCC_solver:
         QCC-VQE minimization is performed."""
 
         # initialize quantities; compute the QMF energy and set this as eqcc_old
-        sim = Simulator()
+        sim = get_backend()
         self.qmf_energy = sim.get_expectation_value(self.qcc_ansatz.qubit_ham, self.qcc_ansatz.qmf_circuit)
         e_qcc, eqcc_old, delta_eqcc = 0., self.qmf_energy, self.deqcc_thresh