+++ tangelo/helpers/utils.py
@@ -12,10 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""This file provides information about what the backends currently installed in
-the user's environment, for the purpose of running / skipping tests, and setting
-a default simulator.
-"""
+""" This file provides miscellaneous utility functions and sets up variables to facilitate testing. """
 
 import functools
 import os
@@ -34,6 +31,17 @@ class HiddenPrints:
     def __exit__(self, exc_type, exc_val, exc_tb):
         sys.stdout.close()
         sys.stdout = self._original_stdout
+
+
+def assert_freq_dict_almost_equal(d1, d2, atol):
+    """ Utility function to check whether two frequency dictionaries are almost equal, for arbitrary tolerance """
+    if d1.keys() != d2.keys():
+        raise AssertionError("Dictionary keys differ. Frequency dictionaries are not almost equal.\n"
+                             f"d1 keys: {d1.keys()} \nd2 keys: {d2.keys()}")
+    else:
+        for k in d1.keys():
+            if abs(d1[k] - d2[k]) > atol:
+                raise AssertionError(f"Frequency {k}, difference above tolerance {atol}: {d1[k]} != {d2[k]}")
 
 
 def is_package_installed(package_name):
@@ -64,7 +72,7 @@ def deprecated(custom_message=""):
     return deprecated_wrapper
 
 
-# List all backends and statevector ones supported by Simulator class
+# List all built-in backends supported
 all_backends = {"qulacs", "qiskit", "cirq", "braket", "projectq", "qdk"}
 all_backends_simulator = {"qulacs", "qiskit", "cirq", "qdk"}
 sv_backends_simulator = {"qulacs", "qiskit", "cirq"}
@@ -73,10 +81,10 @@ sv_backends_simulator = {"qulacs", "qiskit", "cirq"}
 packages = {p: p for p in all_backends}
 packages["qdk"] = "qsharp"
 
-# Figure out what is install in user's environment
+# Figure out what is installed in user's environment
 installed_backends = {p_id for p_id, p_name in packages.items() if is_package_installed(p_name)}
 installed_simulator = installed_backends & all_backends_simulator
 installed_sv_simulator = installed_backends & sv_backends_simulator
 
-# Check if qulacs installed (best performance for tests). If not, defaults to cirq (always installed with openfermion)
+# Check if qulacs installed (better performance for tests). If not, defaults to cirq (always installed with openfermion)
 default_simulator = "qulacs" if "qulacs" in installed_sv_simulator else "cirq"